Add authentication

+++ routes/api.php
@@ -14,13 +14,12 @@ Route::post('/login', [AuthController::class, 'login']);
 Route::post('/register', [AuthController::class, 'register']);
 
 //To do apis
-Route::get('/todo/{id}', [TodoController::class, 'getList']);
-Route::post('/todo/{id}', [TodoController::class, 'addTodo']);
-Route::put('/todoUpdate/{id}', [TodoController::class, 'updateTodo']);
-Route::delete('/todoDelete/{id}', [TodoController::class, 'deleteTodo']);
+Route::get('/todo/{id}', [TodoController::class, 'getList'])->middleware('auth:sanctum');
+Route::post('/todo/{id}', [TodoController::class, 'addTodo'])->middleware('auth:sanctum');
+Route::put('/todoUpdate/{id}', [TodoController::class, 'updateTodo'])->middleware('auth:sanctum');
+Route::delete('/todoDelete/{id}', [TodoController::class, 'deleteTodo'])->middleware('auth:sanctum');
 
 //User apis
-Route::get('/userall', [UserController::class, 'getList']);
-Route::put('/userUpdate/{id}', [UserController::class, 'updateUser']);
-Route::delete('/userDelete/{id}', [UserController::class, 'deleteUser']);
-
+Route::get('/userall', [UserController::class, 'getList'])->middleware('auth:sanctum');
+Route::put('/userUpdate/{id}', [UserController::class, 'updateUser'])->middleware('auth:sanctum');
+Route::delete('/userDelete/{id}', [UserController::class, 'deleteUser'])->middleware('auth:sanctum');
